fix: Default console settings for Python and CodeLLDB
Changed `console` attribute for both the python and codelldb debug
adapters to use the `integratedTerminal` setting. This is a more
sensible default, as without it tools like `dapui` will assume the REPL
should be used - which does not allow for user input.

I have left the other adapters untouched, as I do not have a means of
testing the changes right now and have only experienced the problem with
Rust (CodeLLDB) and Python.